+++ lib/cerner/oauth1a/access_token_agent.rb
@@ -25,6 +25,8 @@ module Cerner
       attr_reader :consumer_key
       # Returns the String Consumer Secret.
       attr_reader :consumer_secret
+      # Returns the String Protection Realm. The realm is root of the access_token_url (scheme + hostname).
+      attr_reader :realm
 
       # Public: Constructs an instance of the agent.
       #
@@ -72,6 +74,7 @@ module Cerner
         @consumer_secret = consumer_secret
 
         @access_token_url = convert_to_http_uri(access_token_url)
+        @realm = canonical_root_url_for(@access_token_url)
 
         @open_timeout = (open_timeout ? open_timeout.to_i : 5)
         @read_timeout = (read_timeout ? read_timeout.to_i : 5)
@@ -215,6 +218,18 @@ module Cerner
         uri
       end
 
+      # Internal: Returns a String containing the canonical root url.
+      #
+      # url - A URL to get the canonical root url String from.
+      #
+      # raises ArgumentError if url is nil.
+      def canonical_root_url_for(url)
+        raise ArgumentError, 'url is nil' unless url
+
+        realm = URI("#{url.scheme}://#{url.host}:#{url.port}")
+        realm.to_s
+      end
+
       # Internal: Prepare a request for #retrieve
       def retrieve_prepare_request(timestamp, nonce, accessor_secret, principal)
         # construct a POST request
@@ -252,14 +267,15 @@ module Cerner
             nonce: nonce,
             timestamp: timestamp,
             token: tuples[:oauth_token],
-            token_secret: tuples[:oauth_token_secret]
+            token_secret: tuples[:oauth_token_secret],
+            realm: @realm
           )
           access_token
         else
           # Extract any OAuth Problems reported in the response
           oauth_data = Protocol.parse_authorization_header(response['WWW-Authenticate'])
           # Raise an error for a failure to acquire a token
-          raise OAuthError.new('unable to acquire token', response.code, oauth_data[:oauth_problem])
+          raise OAuthError.new('unable to acquire token', response.code, oauth_data[:oauth_problem], nil, @realm)
         end
       end
 
@@ -278,10 +294,10 @@ module Cerner
         when Net::HTTPSuccess
           parsed_response = JSON.parse(response.body)
           aes_key = parsed_response.dig('aesKey', 'secretKey')
-          raise OAuthError, 'AES secret key retrieved was invalid' unless aes_key
+          raise OAuthError.new('AES secret key retrieved was invalid', nil, nil, nil, @realm) unless aes_key
 
           rsa_key = parsed_response.dig('rsaKey', 'publicKey')
-          raise OAuthError, 'RSA public key retrieved was invalid' unless rsa_key
+          raise OAuthError.new('RSA public key retrieved was invalid', nil, nil, nil, @realm) unless rsa_key
 
           Keys.new(
             version: keys_version,
@@ -292,7 +308,7 @@ module Cerner
           # Extract any OAuth Problems reported in the response
           oauth_data = Protocol.parse_authorization_header(response['WWW-Authenticate'])
           # Raise an error for a failure to acquire keys
-          raise OAuthError.new('unable to acquire keys', response.code, oauth_data[:oauth_problem])
+          raise OAuthError.new('unable to acquire keys', response.code, oauth_data[:oauth_problem], nil, @realm)
         end
       end
     end